requirements.txt
asgiref==3.8.1
attrs==23.2.0
autobahn
Automat==22.10.0
cffi==1.16.0
celery==5.4.0
channels==4.1.0
channels-redis==4.2.0
constantly==23.10.4
cryptography==42.0.7
daphne==4.1.2
Django
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-redis==5.4.0
django-cors-headers==4.3.1
django-restframework==0.0.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djangorestframework-gis==1.0
drf-yasg==1.21.7
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
inflection==0.5.1
msgpack==1.0.8
packaging==24.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyOpenSSL==24.1.0
python-decouple==3.8
pylint-django==2.5.5
pytz==2024.1
PyYAML==6.0.1
redis==5.0.4
service-identity==24.1.0
setuptools==69.5.1
six==1.16.0
sqlparse==0.5.0
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.11.0
uritemplate==4.1.1
zope.interface==6.4
mypy==1.10.0
mypy-extensions==1.0.0
